+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Apply clippy lint suggestions on macro rules
 - Bump markdownlint-cli2 action from 7 to 8 (<https://github.com/badouralix/kubekeeper/pull/3>)
+- Update magic to detect native kubectl commands (<https://github.com/badouralix/kubekeeper/pull/5>)
 
 ## v2.4.0